First Commit
We up and running! Sort of. These are the first files for the li0n
language. A python interpreter, which reads test.lang and “translates”.
How to run it: Open terminal to the location of the files, and run
“python basic.py test.lang”, and it will output to console. Currently
there is code for print, using the syntax pReach (case sensitive), but
the string does not print correctly at the moment.