Minor changes
In python script, added a few lines that I was going to use to try to
deal with numbers. The way I did it seems really inefficient, might
change it later. Regardless, it’s currently a syntax error. Comment
lines 26-28 if you want the script to run. In test.lang, I added a
number statement to test if the python could handle numbers. Spoiler:
it can’t right now.