First Commit to Branch
basic.py has many comments added but is not yet fully commented. Part
of the number-recognition stuff is in, some is commented out, and it
doesn’t currently work. Just wanted to branch this in case it ends up
taking longer than expected, and so I can quickly get these files
sync’d to my desktop