Simplified parse function!
I added a new function, doPrint, which strips the beginning of tokens.
For example, putting in the token STRING:”Hello World” would output
just Hello World. By using doPrint instead of print in the parse
function, the code is simplified. This will also make future additions
much easier, as you won’t have to deal with the indices of lists as
much.